refactor: separation of the form loader and submission script
