+++ Frontend/resourceRegistrationForm.html
@@ -10,10 +10,11 @@
 
 <body>
     <div class="form-container">
-        <form id="resourceForm">
+        <form id="resourceCreationFormContainer">
              <h1>Formulaire d'Enregistrement des Ressources</h1>
         </form>
     </div>
+    <script src="assets/js/formLoader.js"></script>
     <script src="assets/js/resourceRegistrationForm.js"></script>
     <script src="assets/js/jquery-3.7.1.slim.min.js"></script>
     <script src="assets/js/main.js"></script>
